Assign 4: Like component solution
In the component we name the inputs associated with the component
fields. These are bound to the app component tweet property in the
app.component.html (inside the likes element).

In the likes component html, we use class binding inside the
glyphicon-heart span to add the custom highlighted class when
the component isActive field is true.

The (click) action is bound to the onClick method in the component.